ai-gen/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
jsonschema==4.19.2
python-dotenv==1.0.0
pytest==7.4.3
google-generativeai==0.8.3
openai>=1.30.0
psycopg2-binary==2.9.9
